model/requirements.txt
rasterio==1.1.1
matplotlib
pandas
landsatxplore
sklearn
Pillow
affine
xmltodict
pyproj
earthpy
pickle
